--- src/components/Header/AutoComplete/index.js
@@ -1,1 +1,0 @@
-export { default as AutoComplete } from './AutoComplete.react';
